create own function for revisit and create feedback for back button if no list is uploaded
